Use rgl 0.5's new bipartite_sets function
It's not a perfect replacement for my existing implementation,
especially concerning empty graphs, but it did reveal a bug
in the case of a graph with a single vertex.

Surprisingly, after this change the performance "curve" no longer
shows that dramatic difference between odd and even size graphs.
Both phases show the same, poorer performance.